fixed confirm bug

+++ app/controllers/list.js
@@ -4,7 +4,6 @@ export default Ember.Controller.extend({
   isShowingListModal: false,
   isShowingItemModal: false,
   actions: {
-
     createList: function() {
       var list = this.store.createRecord('list', { title: '', description: ''});
       var item = this.store.createRecord('item', { title: '', description: '', list: list});
@@ -15,10 +14,10 @@ export default Ember.Controller.extend({
       this.store.createRecord('item', { title: '', description: '', list: list});
     },
     deleteList: function(list) {
-      confirm("Are you sure you want to delete this list?");
-      list.deleteRecord();
-      item.destroyRecord();
-
+      var c = confirm("Are you sure you want to delete this list?");
+      if (c == true) {
+        list.deleteRecord();
+      }
     },
     toggleListModal: function() {
       this.toggleProperty('isShowingListModal');
